updates to lab 8
note that all other components are to be checked. it seems as if dat_mem is working as expected when given the correct inputs and outputs. Now, all other components must be checked to make sure that I can make the entire thing work.

I also need to make sure that I understand what each of the signals is doing so I can go through and make notations so I can operate the rest correctly when it comes time to upload to the board.