WORK
so want to be able to see alu things and rs, rt data expectations,

try outputing the register number as a default to test as well

take a look at the control more incase it needs to function.

need to add more signals to verify that the output from the instruction sis correct and that the component are correct.